requirements.txt
streamlit
google-generativeai
langchain
langchain-community
pypdf
chromadb
langchain-google-genai
pillow